requirements.txt
pyyaml
python-ldap
requests
